+++ src/Wt/Dbo/backend/Sqlite3.C
@@ -73,7 +73,7 @@ public:
     : db_(db),
       sql_(sql)
   {
-    LOG_DEBUG("{} for: {}", this, sql_);
+    LOG_DEBUG("{} for: {}", (long long)this, sql_);
 
 #if SQLITE_VERSION_NUMBER >= 3003009
     int err = sqlite3_prepare_v2(db_.connection(), sql.c_str(),
@@ -110,7 +110,7 @@ public:
 
   virtual void bind(int column, const std::string& value) override
   {
-    LOG_DEBUG("{} bind {} {}", this, column, value);
+    LOG_DEBUG("{} bind {} {}", (long long)this, column, value);
 
     int err = sqlite3_bind_text(st_, column + 1, value.c_str(),
 				static_cast<int>(value.length()),
@@ -121,7 +121,7 @@ public:
 
   virtual void bind(int column, short value) override
   {
-    LOG_DEBUG("{} bind {} {}", this, column, value);
+    LOG_DEBUG("{} bind {} {}", (long long)this, column, value);
 
     int err = sqlite3_bind_int(st_, column + 1, value);
 
@@ -130,7 +130,7 @@ public:
 
   virtual void bind(int column, int value) override
   {
-    LOG_DEBUG("{} bind {} {}", this, column, value);
+    LOG_DEBUG("{} bind {} {}", (long long)this, column, value);
 
     int err = sqlite3_bind_int(st_, column + 1, value);
 
@@ -139,7 +139,7 @@ public:
 
   virtual void bind(int column, long long value) override
   {
-    LOG_DEBUG("{} bind {} {}", this, column, value);
+    LOG_DEBUG("{} bind {} {}", (long long)this, column, value);
 
     int err = sqlite3_bind_int64(st_, column + 1, value);
 
@@ -153,7 +153,7 @@ public:
 
   virtual void bind(int column, double value) override
   {
-    LOG_DEBUG("{} bind {} {}", this, column, value);
+    LOG_DEBUG("{} bind {} {}", (long long)this, column, value);
 
     int err;
     if (isNaN(value))
@@ -166,7 +166,7 @@ public:
 
   virtual void bind(int column, const std::chrono::duration<int, std::milli> & value) override
   {
-    LOG_DEBUG("{} bind {} {}ms", this, column, value.count());
+    LOG_DEBUG("{} bind {} {}ms", (long long)this, column, value.count());
 
     long long msec = value.count();
     int err = sqlite3_bind_int64(st_, column + 1, msec);
@@ -240,7 +240,7 @@ public:
 
   virtual void bind(int column, const std::vector<unsigned char>& value) override
   {
-    LOG_DEBUG("{} bind {} (blob, size={})", this, column, value.size());
+    LOG_DEBUG("{} bind {} (blob, size={})", (long long)this, column, value.size());
 
     int err;
 
@@ -255,7 +255,7 @@ public:
 
   virtual void bindNull(int column) override
   {
-    LOG_DEBUG("{} bind {} null", this, column);
+    LOG_DEBUG("{} bind {} null", (long long)this, column);
 
     int err = sqlite3_bind_null(st_, column + 1);
 
@@ -335,7 +335,7 @@ public:
 
     *value = (const char *)sqlite3_column_text(st_, column);
 
-    LOG_DEBUG("{} result string {} {}", this, column, value);
+    LOG_DEBUG("{} result string {} {}", (long long)this, column, value);
 
     return true;
   }
@@ -358,7 +358,7 @@ public:
     *value = 42;
     *value = sqlite3_column_int(st_, column);
 
-    LOG_DEBUG("{} result int {} {}", this, column, value);
+    LOG_DEBUG("{} result int {} {}", (long long)this, column, value);
 
     return true;
   }
@@ -370,7 +370,7 @@ public:
 
     *value = sqlite3_column_int64(st_, column);
 
-    LOG_DEBUG("{} result long long {} {}", this, column, value);
+    LOG_DEBUG("{} result long long {} {}", (long long)this, column, value);
 
     return true;
   }
@@ -388,7 +388,7 @@ public:
       }
     }
 
-    LOG_DEBUG("{} result float {} {}", this, column, value);
+    LOG_DEBUG("{} result float {} {}", (long long)this, column, value);
 
     return true;
   }
@@ -406,7 +406,7 @@ public:
       }
     }
 
-    LOG_DEBUG("{} result double {} {}", this, column, value);
+    LOG_DEBUG("{} result double {} {}", (long long)this, column, value);
 
     return true;
   }
@@ -420,7 +420,7 @@ public:
 
     *value = std::chrono::milliseconds(msec);
 
-    LOG_DEBUG("{} result time_duration {} {}ms", this, column, value->count());
+    LOG_DEBUG("{} result time_duration {} {}ms", (long long)this, column, value->count());
 
     return true;
   }
@@ -535,7 +535,7 @@ public:
     value->resize(s);
     std::copy(v, v + s, value->begin());
 
-    LOG_DEBUG("{} result blob {} (blob, size = {})", this, column, s);
+    LOG_DEBUG("{} result blob {} (blob, size = {})", (long long)this, column, s);
 
     return true;
   }
